Replace Vercel API function with Netlify function for form processing
- Remove Vercel-specific api/process-form-submissions.ts file
- Create equivalent Netlify function in netlify/functions/
- Add @netlify/functions dependency for TypeScript types
- Update documentation (GOOGLE_SHEETS_SETUP_GUIDE.md, FORM_INTEGRATION_SUMMARY.md)
- Update all API endpoint URLs from /api/* to /.netlify/functions/*
- Add missing Google Sheets environment variables to .env.example
