+++ .env.example
@@ -7,6 +7,8 @@ VITE_GEMINI_API_KEY=
 # Google Sheets Form Integration
 GOOGLE_SHEETS_CLIENT_EMAIL=your-service-account@project.iam.gserviceaccount.com
 GOOGLE_SHEETS_PRIVATE_KEY="-----BEGIN PRIVATE KEY-----\nYour private key here\n-----END PRIVATE KEY-----\n"
+GOOGLE_SHEETS_SPREADSHEET_ID=1MHSmxDUdTrc9SIBDCl_hiHNxqoOHMv19nHeMN1_MVUo
+GOOGLE_SHEETS_SHEET_NAME=Form Responses 1
 FORM_PROCESSING_SECRET=your-random-secret-key-generate-with-openssl
 
 # Email Service (Resend)
