+++ package.json
@@ -60,6 +60,7 @@
     "vaul": "^1.1.2"
   },
   "devDependencies": {
+    "@netlify/functions": "^5.1.2",
     "@types/node": "^20.10.0",
     "@vitejs/plugin-react-swc": "^3.10.2",
     "vite": "^7.3.1"
